requirements.txt
# Neural Pathway Routing PoC - Dependencies
# Python 3.10+

# Core ML frameworks
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0

# Transformers and training
transformers==4.37.0
peft==0.8.0
accelerate==0.26.0
datasets==2.16.0
tokenizers==0.15.0

# API integrations
anthropic==0.18.0
openai==1.12.0

# Experiment tracking
wandb==0.16.2
tensorboard==2.15.1

# Scientific computing
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0
umap-learn==0.5.5

# Data handling
pandas==2.1.4
pyarrow==14.0.2

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Utilities
tqdm==4.66.1
click==8.1.7
rich==13.7.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
